for tuning we will no longer tune the drive correction coefficients separately. Instead we will use the turn coefficients but without the feedforward term. We added a helper method to the feedforward class that enables us to access only the feedback coefficients.
We changed the way exit conditions are to be handled for drive and turn.  In addition to checking for stability and position tolerance, we have a separate exit condition for checking if the robot is moving at all using an epsilon equals. The goal of this is to eliminate situations where the robot does not meet its steady state error requirements and will not be able to complete these requirements. Generally this will only ever occur in situations where the robot is effected by friction and as a result will still be NEAR the setpoint.

Adjustments to tuning.